fix create symlink failed of file exists when install package
When install packages such as python-argparse-manpage-help, The error
info is "Failed to create symlink from ... to ..."
When createing a soft link for the interpreter, if there is already
a broken soft link, it cannot be created properly. Therefore, delete
the broken soft link first, and then create a new one.