fix install multiple pkg with dependencies and failed to remove
when install packages with dependencies and then remove, such as:
1. epkg install groff-help groff groff-base
2. epkg remove groff
The return info is: "No packages to remove". but the groff can be
seen in "epkg list groff".
When multiple packages are input, their depends_depth are initialy
set to 0. and since they do not analyze dependencies between them,
the depends_depth is not updated. When removed, the packages are
treated as independent packages and are keeped.