environment: use local channel_config on epkg init
The create_environment() function was calling crate::models::channel_config().format.clone()
which tries to load the global channel configuration. During epkg init, the global config
is set to "base" environment, but when creating the "main" environment, this caused a
mismatch where it tried to read /root/.epkg/config/envs/base.yaml instead of the correct
environment config.

Changed the function to use the local channel_config variable that was just created and
saved to env_channel_yaml, ensuring the correct format is used for environment creation.

Created register_environment_for(name, env_config) function that takes environment config
as parameter, and refactored register_environment() to use it. This fixes the issue where
register_environment() was calling crate::models::env_config() which tries to load the
global environment config.

Updated light_init() to load the environment config after creating it and pass it to
register_environment_for() to ensure the correct environment config is used.

Fixes:
- epkg init crashes with "Failed to deserialize env config: Failed to read file: /root/.epkg/config/envs/base.yaml"
- occurs when creating 'main' environment but trying to read 'base' environment config
- happens because global config.common.env is set to "base" during entire init process

Signed-off-by: Wu Fengguang <wfg@mail.ustc.edu.cn>