utils: fix create_common_symlinks() checking target existence in host rootfs
The function was checking if symlink targets exist in the host rootfs
(Path::new("/")) instead of the environment rootfs (env_root). This caused
dead symlinks to be created when targets existed on the host but not in the
environment.

Changed target_path construction to check within env_root:
- For absolute targets (starting with "/"), check env_root.join(&target[1..])
- For relative targets, check relative to symlink's parent within env_root

Also changed symlink() call to use the original target string instead of
target_path, so symlinks point to relative or absolute paths as originally
specified.

Fixes:
- Dead symlink ~/.epkg/envs/alpine/usr/bin/sh -> /bin/bash (target doesn't exist in env)
- Dead symlink ~/.epkg/envs/alpine/usr/bin/awk -> /usr/bin/mawk (target doesn't exist in env)

Signed-off-by: Wu Fengguang <wfg@mail.ustc.edu.cn>