env: make ENV_NAME optional in export subcommand
Changed the export subcommand to accept ENV_NAME as an optional positional
argument instead of required. When ENV_NAME is not provided, the command
now uses the environment specified via the global -e flag (stored in
config().common.env).

Updated command_env() handler to use config().common.env as fallback when
ENV_NAME is not provided in sub_matches, allowing the export command to
work with both forms:
- epkg -e ENV_NAME env export --output FILE
- epkg env export ENV_NAME --output FILE

Fixes:
- epkg -e test-export env export --output /tmp/epkg-export-test-export.yaml
  error: one or more required arguments were not provided
  ERROR: epkg command failed with exit code 2

Signed-off-by: Wu Fengguang <wfg@mail.ustc.edu.cn>