remove: check target path type before unlinking files
The unlink_package() function was checking if the source file was a
directory but not checking if the target path in the environment was a
directory. When attempting to remove a file that exists as a directory
in the target environment, fs::remove_file() would fail with "Is a
directory" error.

Now the code checks target_metadata.is_dir() before attempting to remove
the target path. If the target is a directory, it is skipped (directories
are typically shared and shouldn't be removed during package removal).
Only files and symlinks are removed.

Fixes:
   0: Failed to unlink package filesystem__2025.10.12-1__any
   1: Is a directory (os error 21)
Location:
   src/remove.rs:38
   Removing package file: /home/wfg/.epkg/envs/archlinux/usr/sbin

Also protect symlinks by create_environment_directories().
Fixes archlinux /lib32/ld-linux-x86-64.so.2: No such file or directory

Signed-off-by: Wu Fengguang <wfg@mail.ustc.edu.cn>