run: change working directory to / instead of env_root
When `chdir_to_env_root` is enabled, we now change to `/` instead of
`env_root` because the `/opt/epkg/envs/` directory can be empty when
the environment is standalone mounted. After namespace setup and bind
mounts, the root filesystem already contains the environment contents,
so changing to `/` achieves the same effect while avoiding the error
when `env_root` doesn't exist as a directory.

Fixes:
- Failed to change to environment root directory /opt/epkg/envs/epkgtest_a/puba: No such file or directory (os error 2)
- Command '/opt/epkg/envs/epkgtest_a/puba/usr/bin/ldconfig' exited with code 1 (no_exit=true, continuing)

Signed-off-by: Wu Fengguang <wfg@mail.ustc.edu.cn>